リーグリストとチームリストをコンポーネントに分けた
この2つをそれぞれコンポーネントに分ける
